pyproject.toml
[tool.poetry]
name = "samesport-bot"
version = "0.1.0"
description = ""
authors = ["kbondar17 <ironbondar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.19"
SQLAlchemy = "^1.4.31"
python-dotenv = "^0.19.2"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
